requirements.txt
fastapi
uvicorn
python-multipart
pillow
reportlab
Jinja2
opencv-python
numpy
pytesseract
